backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.30.0",
    "bcrypt>=4.2.0",
    "email-validator>=2.2.0",
    "fastapi>=0.135.3",
    "psycopg[binary]>=3.2.3",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.9.0",
    "python-multipart>=0.0.12",
    "ruff>=0.15.9",
    "sqlalchemy[asyncio]>=2.0.49",
    "uvicorn>=0.44.0",
]
